+++ service/server-common/src/main/java/com/hansun/dto/Province.java
@@ -4,14 +4,14 @@ package com.hansun.dto;
  * Created by yuanl2 on 2017/3/29.
  */
 public class Province {
-    private int id;
+    private short id;
     private String name;
 
-    public int getId() {
+    public short getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(short id) {
         this.id = id;
     }
