+++ service/server-common/src/main/java/com/hansun/server/common/DeviceStatus.java
@@ -4,11 +4,11 @@ package com.hansun.server.common;
  * Created by yuanl2 on 2017/3/31.
  */
 public class DeviceStatus {
-    public static final int DISCONNECTED = 0;
-    public static final int IDLE = 1;
-    public static final int INVALID = 2;
-    public static final int STARTTASK = 3;
-    public static final int SERVICE = 4;
-    public static final int BADNETWORK =5;
-    public static final int FAULT = 6;
+    public static final byte DISCONNECTED = 0;
+    public static final byte IDLE = 1;
+    public static final byte INVALID = 2;
+    public static final byte STARTTASK = 3;
+    public static final byte SERVICE = 4;
+    public static final byte BADNETWORK =5;
+    public static final byte FAULT = 6;
 }
