+++ service/server-common/src/main/java/com/hansun/server/common/OrderStatus.java
@@ -5,12 +5,12 @@ package com.hansun.server.common;
  */
 public class OrderStatus {
 
-    public static final int CREATED = 0;   //未支付
-    public static final int NOTSTART = 1;     //已支付未启动
-    public static final int PAYDONE = 2;     //支付通知成功
-    public static final int SERVICE = 3;  //已支付运行中
-    public static final int FINISH = 4;    //已支付已完成
-    public static final int DEVICE_ERROR = 5;
-    public static final int USER_NOT_PAY = 6;
-    public static final int UNKNOW = 7;
+    public static final short CREATED = 0;   //未支付
+    public static final short NOTSTART = 1;     //已支付未启动
+    public static final short PAYDONE = 2;     //支付通知成功
+    public static final short SERVICE = 3;  //已支付运行中
+    public static final short FINISH = 4;    //已支付已完成
+    public static final short DEVICE_ERROR = 5;
+    public static final short USER_NOT_PAY = 6;
+    public static final short UNKNOW = 7;
 }
