+++ service/server/src/main/java/com/hansun/server/commu/SocketHandler4G.java
@@ -29,7 +29,7 @@ public class SocketHandler4G extends AbstractHandler implements IHandler {
 
     private final Logger logger = LoggerFactory.getLogger(getClass());
 
-    private Map<Integer, Integer> portStatus = new ConcurrentHashMap<>();
+    private Map<Integer, Byte> portStatus = new ConcurrentHashMap<>();
 
     private String deviceName;
 
@@ -221,7 +221,7 @@ public class SocketHandler4G extends AbstractHandler implements IHandler {
         this.hasConnected = hasConnected;
     }
 
-    public Map<Integer, Integer> getPortStatus() {
+    public Map<Integer, Byte> getPortStatus() {
         return portStatus;
     }
